requirements.txt
absl-py==2.1.0
alembic==1.13.3
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.36.0
DateTime==5.5
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
docker==7.1.0
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==2.17.1
mlflow-skinny==2.17.1
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.5.8
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optree==0.13.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
patsy==0.5.6
pillow==10.4.0
platformdirs==4.1.0
plotly==5.24.1
-e git+https://github.com/minalmmm/Project-6--Pharmaceutical-Sales-prediction-across-multiple-stores.git@62864eae8a1955d7cd4e8c72e9b829d3ee5cdb98#egg=project_6
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==5.9.7
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.8.2
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
waitress==3.0.0
wcwidth==0.2.12
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
xgboost==2.1.2
zipp==3.20.2
zope.interface==7.0.3
streamlit
pandas
numpy
matplotlib
tensorflow
mlflow